Fix resolver thrown-error-propagation
The subscription for graphql calls was producing a promise that didn’t have errors caught. This meant that when an apollo-link-state resolver threw an error, that error was uncatchable in Jest, and that means that users can’t validate API failures when they occur.

This catches errors thrown in resolvers, and funnels them into the observerErrorHandler.